+++ test/test_daq_mqtt.py
@@ -44,6 +44,40 @@ def test_mqtt_to_influxdb_json_single(machinery, create_influxdb, reset_influxdb
 
 @pytest_twisted.inlineCallbacks
 @pytest.mark.mqtt
+def test_mqtt_to_influxdb_json_bulk(machinery, create_influxdb, reset_influxdb):
+    """
+    Publish multiple readings in JSON format to MQTT broker
+    and proof it is stored in the InfluxDB database.
+    """
+
+    # Submit multiple measurements, without timestamp.
+    data = [
+        {
+            'temperature': 21.42,
+            'humidity': 41.55,
+        },
+        {
+            'temperature': 42.84,
+            'humidity': 83.1,
+        },
+    ]
+    yield threads.deferToThread(mqtt_json_sensor, settings.mqtt_topic_json, data)
+
+    # Wait for some time to process the message.
+    yield sleep(PROCESS_DELAY_MQTT)
+
+    # Proof that data arrived in InfluxDB.
+    record = influx_sensors.get_record(index=0)
+    del record['time']
+    assert record == {u'temperature': 21.42, u'humidity': 41.55}
+
+    record = influx_sensors.get_record(index=1)
+    del record['time']
+    assert record == {u'temperature': 42.84, u'humidity': 83.1}
+
+
+@pytest_twisted.inlineCallbacks
+@pytest.mark.mqtt
 @pytest.mark.legacy
 def test_mqtt_to_influxdb_json_legacy_topic(machinery, create_influxdb, reset_influxdb):
     """